users/me endpoint. Authorization. Config file with env variables. requirements.in file, refactored schemes.
