+++ app/routers/users.py
@@ -3,9 +3,15 @@ from datetime import timedelta
 from typing import Optional, List
 
 from fastapi import APIRouter, Depends, HTTPException, status
+from fastapi.security import OAuth2PasswordRequestForm
 from sqlalchemy.orm import Session
 
-from app.auth import ACCESS_TOKEN_EXPIRE_MINUTES, create_access_token, authenticate_user
+from app.auth import (
+    create_access_token,
+    authenticate_user,
+    get_current_active_user,
+)
+from app.config import ACCESS_TOKEN_EXPIRE_MINUTES
 from app.crud.crud_user import (
     get_user_by_email,
     get_user_list,
@@ -13,7 +19,8 @@ from app.crud.crud_user import (
     get_user_by_id,
 )
 from app.deps import get_db
-from app.schemas.user import User, CreateUser, UserCredentials
+from app.schemas.auth import Token
+from app.schemas.user import User, CreateUser
 
 router = APIRouter(
     prefix="/users",
@@ -38,35 +45,33 @@ def create_user(user_data: CreateUser, db: Session = Depends(get_db)):
     return new_user
 
 
-@router.get("/{email}", response_model=User)
-def get_user(email: str, db: Session = Depends(get_db)):
-    user = get_user_by_email(db, email)
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"User with email {email} not found.",
-        )
-    return user
+# @router.get("/{email}", response_model=User)
+# def get_user(email: str, db: Session = Depends(get_db)):
+#     user = get_user_by_email(db, email)
+#     if not user:
+#         raise HTTPException(
+#             status_code=status.HTTP_404_NOT_FOUND,
+#             detail=f"User with email {email} not found.",
+#         )
+#     return user
+#
+#
+# @router.get("/{user_id}", response_model=User)
+# def get_user(user_id: uuid.UUID, db: Session = Depends(get_db)):
+#     user = get_user_by_id(db, user_id)
+#     if not user:
+#         raise HTTPException(
+#             status_code=status.HTTP_404_NOT_FOUND,
+#             detail=f"User with id {user_id} not found.",
+#         )
+#     return user
 
 
-@router.get("/{user_id}", response_model=User)
-def get_user(user_id: uuid.UUID, db: Session = Depends(get_db)):
-    user = get_user_by_id(db, user_id)
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"User with id {user_id} not found.",
-        )
-    return user
-
-
-@router.post("/token")
+@router.post("/token", response_model=Token)
 def login_for_access_token(
-    user_credentials: UserCredentials, db: Session = Depends(get_db)
+    form_data: OAuth2PasswordRequestForm = Depends(), db: Session = Depends(get_db)
 ):
-    user = authenticate_user(
-        db, email=user_credentials.email, password=user_credentials.password
-    )
+    user = authenticate_user(db, email=form_data.username, password=form_data.password)
     if not user:
         raise HTTPException(
             status_code=status.HTTP_401_UNAUTHORIZED,
@@ -78,3 +83,8 @@ def login_for_access_token(
         data={"sub": user.email}, expires_delta=access_token_expires
     )
     return {"access_token": access_token, "token_type": "bearer"}
+
+
+@router.get("/me", response_model=User)
+def read_users_me(current_user: User = Depends(get_current_active_user)):
+    return current_user
