+++ app/schemas/user.py
@@ -9,7 +9,7 @@ class UserBase(BaseModel):
     last_name: Optional[str]
     email: str  # Optional[EmailStr] = None
     # phone_number: str
-    is_active: bool
+    is_active: bool = None
 
 
 class User(UserBase):
@@ -21,19 +21,3 @@ class User(UserBase):
 
 class CreateUser(UserBase):
     password: str
-
-
-# todo: pass as body. better way.
-class UserCredentials(BaseModel):
-    email: str
-    password: str
-
-
-# todo: move to separate file
-class Token(BaseModel):
-    access_token: str
-    token_type: str
-
-
-class TokenData(BaseModel):
-    user_email: Union[str, None] = None
