+++ docker-compose.yml
@@ -24,8 +24,8 @@ services:
       - db
     environment:
       - SQLALCHEMY_DATABASE_URL=postgresql+psycopg2://postgres:password@db:5432/vwire_db
-      - SIGN_SECRET_KEY=e025ce978c17a012241a16ca1457e2d4c4f9d7e8f47c2cd2b5ac629c43c64db4  # openssl rand -hex 32
-      - SIGN_ALGORITHM=HS256
+      - JWT_SECRET_KEY=e025ce978c17a012241a16ca1457e2d4c4f9d7e8f47c2cd2b5ac629c43c64db4  # openssl rand -hex 32
+      - JWT_ALGORITHM=HS256
       - ACCESS_TOKEN_EXPIRE_MINUTES=60
 volumes:
   db_data:
